refactor: move header to header.rs, reuse constants in tempo-zone
Moves ZoneHeader into its own module and replaces duplicated address
and storage slot constants in tempo-zone's abi.rs with re-exports
from zone-primitives.
