+++ crates/tempo-zone/Cargo.toml
@@ -24,6 +24,7 @@ tempo-alloy = { workspace = true, features = ["tempo-compat"] }
 tempo-contracts.workspace = true
 tempo-precompiles.workspace = true
 tempo-precompiles-macros.workspace = true
+zone-primitives.workspace = true
 zone-rpc.workspace = true
 
 # reth
